We should use Deque instead of Stack because it's more sensible in terms of inheritance. The fact that Stack extends Vector is really strange.Early in Java, inheritance was overused IMO - Properties being another example.
Deque exposes a set of operations which is all about being able to fetch/add/remove items from the start or end of a collection, iterate etc - and that's it. There's deliberately no way to access an element by position, which Stack exposes because it's a subclass of Vector.

Oh, and also Stack has no interface, so if you know you need Stack operations you end up committing to a specific concrete class, which isn't usually a good idea.

And earlier r.peek()<temp was sorting the stack in descending order so I had changed it to r.peek()>temp now it is sorting a stack in ascending order